Better Collision
Collision checking while standing and also it checks for a range of X
values, not just an exact match. Added a function to enemy to create a
MinX and MaxX value, based on current x coordinate plus and minus some
range. Then, a function returns the min and max x values for range to
gamePanel, where the player x value can be compared.